chore: refactor dashboard router test
